refactor: rename to useEditableObject and update comments
